pyproject.toml
[tool.poetry]
name = "dicom_service"
version = "0.1.0"
description = "Sync Flask + gRPC DICOM microservice"
authors = ["Niharika Maheshwari niharikamaheshwari@gmail.com"]

packages = [
    { include = "grpc_backend" },
    { include = "rest_api" }
]

[tool.poetry.dependencies]
python = ">=3.12"
flask = "^3.1.1"
grpcio = ">=1.73.1"
pydicom = "^3.0.1"
pillow = "^11.2.1"
protobuf = "^6.31.1"
sqlalchemy = "^2.0.41"
numpy = "^2.3.1"
grpcio-tools = "^1.73.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
